split up forest source files
