+++ src/mesh/forest/relative_orientation.hpp
@@ -1,0 +1,56 @@
+//========================================================================================
+// (C) (or copyright) 2023. Triad National Security, LLC. All rights reserved.
+//
+// This program was produced under U.S. Government contract 89233218CNA000001 for Los
+// Alamos National Laboratory (LANL), which is operated by Triad National Security, LLC
+// for the U.S. Department of Energy/National Nuclear Security Administration. All rights
+// in the program are reserved by Triad National Security, LLC, and the U.S. Department
+// of Energy/National Nuclear Security Administration. The Government is granted for
+// itself and others acting on its behalf a nonexclusive, paid-up, irrevocable worldwide
+// license in this material to reproduce, prepare derivative works, distribute copies to
+// the public, perform publicly and display publicly, and to permit others to do so.
+//========================================================================================
+#ifndef MESH_FOREST_HPP_
+#define MESH_FOREST_HPP_
+
+#include <array>
+#include <map>
+#include <memory>
+#include <set>
+#include <tuple>
+#include <unordered_map>
+#include <unordered_set>
+#include <utility>
+#include <vector>
+
+#include "basic_types.hpp"
+#include "defs.hpp"
+#include "mesh/forest/cell_center_offsets.hpp"
+#include "mesh/logical_location.hpp"
+#include "utils/bit_hacks.hpp"
+#include "utils/indexer.hpp"
+
+namespace parthenon {
+namespace forest {
+
+struct RelativeOrientation {
+  RelativeOrientation() : dir_connection{0, 1, 2}, dir_flip{false, false, false} {};
+
+  void SetDirection(Direction origin, Direction neighbor, bool reversed = false) {
+    dir_connection[static_cast<uint>(origin)] = static_cast<uint>(neighbor);
+    dir_flip[static_cast<uint>(origin)] = reversed;
+  }
+
+  LogicalLocation Transform(const LogicalLocation &loc_in,
+                            std::int64_t destination) const;
+  LogicalLocation TransformBack(const LogicalLocation &loc_in, std::int64_t origin) const;
+
+  bool use_offset = false;
+  std::array<int, 3> offset;
+  std::array<int, 3> dir_connection;
+  std::array<bool, 3> dir_flip;
+};
+} // namespace forest
+} // namespace parthenon
+
+#endif // MESH_FOREST_HPP_
